chore(lint): removed some typescript rules affecting js files
