+++ lib/bootstrap.js
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 const execa = require('execa');
 const WebpackCLI = require('./webpack-cli');
 const { core, commands } = require('./utils/cli-flags');
