+++ lib/commands/external.js
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 // const { prompt } = require('inquirer');
 // const chalk = require('chalk');
 // const logger = require('../utils/logger');
