+++ lib/groups/advanced.js
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 const GroupHelper = require('../utils/group-helper');
 const logger = require('../utils/logger');
