+++ lib/groups/basic.js
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 const GroupHelper = require('../utils/group-helper');
 
 class BasicGroup extends GroupHelper {
