+++ lib/groups/help.js
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 const chalk = require('chalk');
 const commandLineUsage = require('command-line-usage');
