+++ lib/groups/output.js
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 const path = require('path');
 const GroupHelper = require('../utils/group-helper');
