+++ lib/groups/stats.js
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 const GroupHelper = require('../utils/group-helper');
 
 class StatsGroup extends GroupHelper {
