+++ lib/utils/development-config.js
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 const { join, resolve } = require('path');
 const TerserPlugin = require('terser-webpack-plugin');
