+++ lib/utils/group-helper.js
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 const { resolve, join } = require('path');
 const { existsSync } = require('fs');
