+++ lib/utils/interactive.js
@@ -1,5 +1,5 @@
 const chalk = require('chalk');
-const webpack = require('./compiler');
+const webpack = require('./Compiler');
 const ansiEscapes = require('ansi-escapes');
 const readline = require('readline');
