+++ lib/utils/parse-args.js
@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
-
 /**
  * Parse cli args and split these to args for node js and the rest
  *
