+++ lib/utils/process-log.js
@@ -1,6 +1,5 @@
 const logger = require('./logger');
 
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 function logErrorAndExit(error) {
     if (error && error.stack) logger.error(error.stack);
     process.exit(1);
