+++ lib/utils/zero-config.js
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 function getEnvFromOptionsAndMode(mode, optionsObject) {
     const { dev, prod } = optionsObject;
