+++ lib/webpack-cli.js
@@ -1,8 +1,7 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 const { resolve, parse, join } = require('path');
 const { existsSync } = require('fs');
 const GroupHelper = require('./utils/group-helper');
-const Compiler = require('./utils/compiler');
+const { Compiler } = require('./utils/Compiler');
 
 const webpackMerge = require('webpack-merge');
 
@@ -21,6 +20,7 @@ class WebpackCLI extends GroupHelper {
         this.groupMap = new Map();
         this.groups = [];
         this.processingMessageBuffer = [];
+        this.compiler = new Compiler();
     }
     setMappedGroups(args, yargsOptions) {
         const { _all } = args;
@@ -145,15 +145,10 @@ class WebpackCLI extends GroupHelper {
         return groupResult;
     }
 
-    async getCompiler(args, yargsOptions) {
-        const groupResult = await this.processArgs(args, yargsOptions);
-        const { options } = groupResult;
-        return await Compiler.getCompiler(options);
-    }
-
     async run(args, yargsOptions) {
         const groupResult = await this.processArgs(args, yargsOptions);
-        const webpack = await Compiler.webpackInstance(groupResult);
+        this.compiler.createCompiler(groupResult.options);
+        const webpack = await this.compiler.webpackInstance(groupResult);
         return webpack;
     }
