+++ test/config/type/function/webpack.config.js
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 module.exports = () => {
     return {
         entry: './a',
