+++ test/config/type/promise/webpack.config.js
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 module.exports = () => {
     return new Promise(resolve => {
         setTimeout(() => {
